style: font-height, 회원탈퇴 버튼 위치 변경, post description 말줄임 설정 추가
